Emailcheck부분 로직만 구성
controller → usercontroller

domain → entity

- treeuser → user
- treeEnum →

Enum pakage 추가

- EmailType
- ResultType → common.enum안에 resulttype으로
