+++ src/main/java/com/chukapoka/server/common/Enum/EmailType.java
@@ -1,0 +1,9 @@
+package com.chukapoka.server.common.Enum;
+
+public enum EmailType {
+
+    // 등록된 이메일인지 확인 후 이메일타입 확인, 기본값은 Default
+    Default,
+    Google,
+    Kakao
+}
