+++ src/main/java/com/chukapoka/server/common/Enum/ResultType.java
@@ -1,0 +1,7 @@
+package com.chukapoka.server.common.Enum;
+
+public enum ResultType {
+    // 성공, 실패
+    Success,
+    Fail
+}
