--- src/main/java/com/chukapoka/server/treemember/controller/TreeController.java
@@ -1,4 +1,0 @@
-package com.chukapoka.server.treemember.controller;
-
-public class TreeController {
-}
