--- src/main/java/com/chukapoka/server/treemember/controller/TreeItemController.java
@@ -1,4 +1,0 @@
-package com.chukapoka.server.treemember.controller;
-
-public class TreeItemController {
-}
