--- src/main/java/com/chukapoka/server/treemember/domain/ResponseData.java
@@ -1,13 +1,0 @@
-package com.chukapoka.server.treemember.domain;
-
-public class ResponseData {
-
-    public int id;
-    public boolean isSuccess;
-
-    public ResponseData(int id, boolean isSuccess) {
-        this.id = id;
-        this.isSuccess = isSuccess;
-    }
-
-}
