+++ src/main/java/com/chukapoka/server/user/sevice/UserService.java
@@ -1,0 +1,34 @@
+package com.chukapoka.server.user.sevice;
+
+
+
+import com.chukapoka.server.user.dto.*;
+import com.chukapoka.server.user.repository.UserRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+@Service
+public class UserService {
+    @Autowired
+    private UserRepository userRepository;
+
+    /** 이메일 체크 서비스
+     * - 이메일이 등록되어 있는지 확인
+     * - 등록된 이메일이면 로그인, 등록되지 않은 이메일이면 회원가입으로 처리
+     */
+    public EmailCheckResponseDto checkEmail(EmailCheckRequestDto emailCheckRequestDto) {
+        String email = emailCheckRequestDto.getEmail();
+        String emailType = emailCheckRequestDto.getEmailType();
+
+        // 이메일이 이미 등록되어 있는지 확인
+        if (userRepository.existsByEmailAndEmailType(email, emailType)) {
+            // 이메일이 등록되어 있으면 {login, email} 값 반환
+            return new EmailCheckResponseDto("login", email);
+        } else {
+            // 등록되어 있지않다면 회원가입 {join, email} 값 반환
+            return new EmailCheckResponseDto("join", email);
+        }
+    }
+
+
+}
