+++ src/main/resources/application-develop.properties
@@ -1,5 +1,5 @@
 # ?? ?? ??
-server.port=3003
+server.port=3000
 
 # PostgreSQL ?????? ?? ??
 spring.datasource.url=jdbc:postgresql://localhost:5432/tree
